hospital managment system
The Clinic Nallam Management System is a web application that can be entered into using respective  username and password for each individual module. Its database is accessible by the administrator. The data can be retrieved easily.
The data is well protected for personal use and this makes the data processing very fast. This system objectives primarily includes the access of easy& maintainable information.

Clinic Nallam Management System is a CMS based application. CMS stands for Content Management System which is a computer application that supports the creation and modification of digital content. It is often used to support multiple users working in a collaborative environment.